python-server/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain
langchain_google_genai
PyJWT
pymongo
PyPDF2
langchain-community
faiss-cpu
uuid